View interface integration to controller

+++ src/ime/Model/ImageRepository.java
@@ -25,13 +25,13 @@ public interface ImageRepository {
    * Brighten this image by the given brightness constant and return the new image.
    * @param brightnessConstant the constant factor for the image to be brightened by
    */
-  void brightenImage(String imageNameSrc, String imageNameDest, int brightnessConstant);
+  void brightenImage(String imageNameSrc, String imageNameDest, float brightnessConstant);
 
   /**
    * Darken this image by the given darkness constant and return the new image.
    * @param darknessConstant the constant factor for the image to be darkened by
    */
-  void darkenenImage(String imageNameSrc, String imageNameDest, int darknessConstant);
+  void darkenenImage(String imageNameSrc, String imageNameDest, float darknessConstant);
 
   /**
    * Blur the image using appropriate filter and return the new blurred image.
